test: check if process import is result of using default server.ts
